Consistency Fixes
