+++ MonsterTrainModdingAPI/Utilities/AssetBundleUtils.cs
@@ -8,6 +8,18 @@ using System.Reflection;
 
 namespace MonsterTrainModdingAPI.Utilities
 {
+    public class AssetBundleLoadingInfo
+    {
+        public string bundleName;
+        public string assetName;
+        public string pathExtension;
+        public AssetBundleLoadingInfo(string bundleName, string assetName, string pathExtension = "")
+        {
+            this.assetName = assetName;
+            this.bundleName = bundleName;
+            this.pathExtension = pathExtension;
+        }
+    }
     public class AssetBundleUtils
     {
         private static Dictionary<string, AssetBundle> LoadedAssetBundles { get; } = new Dictionary<string, AssetBundle>();
